pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "hr-assistant"
version = "0.1.0"
description = "An ML assistant for HR professionals."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["more-itertools>=10.5.0", "numpy>2", "pandas>=2.2.3"]

[dependency-groups]
docs = [
  "mike",
  "mkdocs",
  "mkdocs-callouts",
  "mkdocs-gen-files",
  "mkdocs-literate-nav",
  "mkdocs-section-index",
  "mkdocs-material",
  "mkdocstrings[python]",
  "mkdocs-backlinks>=0.9.1",
]
train = [
  "folktables>=0.0.12",
  "mlflow>=2.22.0",
  "kaleido==0.2.1",
  "dagster>=1.9.10",
  "dagster-webserver>=1.9.10",
  "dagster-obstore>=0.2.0",
  "sensai-utils>=1.4.0",
  "scikit-learn<1.6",
  "ucimlrepo>=0.0.7",
  "xgboost>=2,<3",
  "optuna>=4.2.1",
  "optuna-integration[sklearn]>=4.2.1",
  "lakefs-spec>=0.11.2",
  "dagster-postgres>=0.26.2",
  "s3fs>=2025.2.0",
  "nannyml>=0.13.0",
  "mlserver>=1.7.0",
  "mlserver-mlflow>=1.7.0",
  "aif360>=0.6.1",
  "shap>=0.47.0",
  "numba>=0.61.0",
  "requests",
  "dagster-graphql>=1.10.15",
  "fairlearn>=0.12.0",
  "dagster-pandas>=0.26.19",
]
deploy = [
  "fastapi>=0.110",
  "mlserver>=1.7.0",
  "mlserver-mlflow>=1.7.0",
  "httpx>=0.28.1",
  "prometheus-fastapi-instrumentator>=7.0.2",
  "psycopg2-binary>=2.9.10",
  "pydantic-settings>=2.8.0",
  "nannyml>=0.13.0",
  "mlserver-sklearn>=1.7.0",
  "matplotlib>=3.10.0",
  "shap>=0.47.0",
]
monitoring = [
  "fastapi>=0.115.8",
  "mlserver>=1.7.0",
  "mlserver-mlflow>=1.7.0",
  "mlserver-sklearn>=1.7.0",
  "nannyml>=0.13.0",
  "psycopg2-binary>=2.9.10",
]
dev = ["ipython>=8.32.0", "pre-commit>=4.2.0", "pytest>=8.3.4", "ruff>=0.9.3"]

[tool.uv.sources]
# Resolve NannyML from Git so we can get the relaxed Jinja2 dependency
# (see https://github.com/NannyML/nannyml/pull/441)
# TODO: Remove this when NannyML releases a new version on PyPI
nannyml = { git = "https://github.com/NannyML/nannyml?ref=main" }

[tool.ruff]
preview = true
exclude = ["openapi_client"]
line-length = 88

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "W",
  "B",
  "UP",
  "C4",
  "PYI",
  "FAST",
  "T10",  # prevent stray debug breakpoints
]
ignore = [
  "E501",    # Line too long
  "RUF029",  # Unused Async (FastAPI routes are async)
  "FAST003", # FastAPI unused path parameter (they can be consumed by dependencies)
  "UP042",   # Allow inheritance from both str and enum.Enum for compatibility
]

[tool.dagster]
module_name = "salary_prediction"
code_location_name = "salary_prediction"
